README/systemd: Set WantedBy to default.target
`multi-user.target` does not work on my raspberry pi, but `default.target` does. I.e. the former results in a systemd service on `inactive(dead)` and the latter results in a systemd unit `active` on startup. My vague understanding is that the former is for headless installations (which my pi is :confused:) and the latter is a catch-all.

Would appreciate if someone could test this on their machine where `multi-user.target` works.